Express result of maximization in terms of original variables.
When options->Maximize is specified, each variable x is replaced
by B - x, with B a big parameter, to be able to solve the maximum
with pip.
This patch reverts this operation after the solution has been
computed.  It also remove the big parameter if it didn't appear
in the input.
The option has been renamed from Max to Maximize to signal
to users of the old behaviour that they should change their code.
This commit also simplifies the tab_Matrix2Tableau code.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>