integrer.c: don't create new parameters that depend on the big parameter
It doesn't make sense for the solution to depend on the residue class
of a parameter that is assumed to be as big as we want.
Instead, we assume that the big parameter is divisible by any number.

The code to remove the big parameter from the solution tacitly
assumed that no new parameters would depend on the new parameter
and so would be incorrect if such parameter did occur.