source/tab.c: align struct A chunk on sizeof(Entier)
tab_alloc constructs a pointer to a struct T by first making room
for a struct A on a pointer returned by malloc.
Some architectures require special alignments on elements of a
given size (say long long on Sparc) and the compiler ensures
that any such element of a structure has the proper alignment
provided the structure itself is properly aligned.
Although malloc will ensure that the pointer it returns is
indeed properly aligned, this will no longer hold if the address
is constructed by adding an arbitrary value to this return value.
By rounding the size of a struct A to a multiple of the size
of an Entier, the alignment will remain correct.