integrer: correct check for available space in the context tableau.
During the `Gomory Cuts' step, the solver introduces new parameters to
the problem and adds one constraint to the problem tableau and two
constraints to the context tableau.
Since rows and columns are allocated in advance to reduce the number
of malloc calls, the solver has to check whether there are enough
rows and columns left in the tableau, but the check for the number
of columns was incorrect.  It would check whether there was enough
room for the _current_ number of parameters (nparam + 1 for the constant)
and not for the new parameter.