exam_coef: also check inequalities not containing the big parameter
If there is a big parameter in the system, then exam_coef would
completely ignore all inequalities not containing the big parameter,
leaving the determination of the sign to the much more expensive
compa_test.  In particular, if the only parameter in the system
is a big parameter, then compa_test should never be called,
but it was being called before.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>