traiter.c: remove needless allocations of discr array
There were to allocations of a discr array.
One in traiter, that was completely useless, and one in exam_coeff
that was only used as a temporary copy of data that is still
available when the temporary copy was used.
Instead, we now directly copy the data from the original source.

These allocation are mostly irrelevant when using fixed size
integers, but they hurt quite a lot when using gmp integers.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>