doc: fix description of how to handle possibly negative parameters
Commit 3b332f7 (doc: clarify handling of negative unknowns and parameters)
"clarified" the original description, but the new description implied
that every possbibly negative parameter requires the introduction of
a separate additional parameter, whereas in fact one additional parameter
is sufficient.  Revert the description to something that is closer
to the original description.

Unfortunately, the implementation in piplib was based on the "clarified"
description and therefore doubles the number of parameters, whereas
only one extra parameter would do.  This should be fixed at some point.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>